+++ app/index.tsx
@@ -2,5 +2,5 @@ import { Redirect } from 'expo-router';
 
 export default function Index() {
   
-  return <Redirect href="/auth" />;
+  return <Redirect href="/login" />;
 }
